pyproject.toml
[tool.poetry]
name = "library-api"
version = "0.1.0"
description = "REST API for library management"
authors = ["Mykyta Lozhevych <m.lozhevych@setuniversity.edu.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
flask-smorest = "^0.45.0"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.1"
cryptography = "^44.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
